// src/types/menuItem.ts
export interface menuItems {
    available: boolean;
    category : string;
    cusine: string;
    description : string;
    foodItemId : string;
    imageURL: string;
    price:number;
    rating : number;
    restaurantId : string;
    serves: number;
    title: string;
    type: string;
    _id: string;

}